add tuples mapping to get data from native query

+++ src/main/java/com/example/chessbackendwithjpa/controller/dto/ScoresByPlayerDTO.java
@@ -1,6 +1,18 @@
 package com.example.chessbackendwithjpa.controller.dto;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Date;
 
-public record ScoresByPlayerDTO(int id, String name, String site, Date start_date, Date end_date, int years, int months, int elo, double points, int games, int performance) {
+public record ScoresByPlayerDTO(Long id,
+                                String name,
+                                String site,
+                                Date start_date,
+                                Date end_date,
+                                BigDecimal years,
+                                BigDecimal months,
+                                BigDecimal elo,
+                                BigDecimal points,
+                                Long games,
+                                BigDecimal performance) {
 }
